+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/room/send/queue/QueuedTask.kt
@@ -23,7 +23,13 @@ abstract class QueuedTask : Cancelable {
 
     private var hasBeenCancelled: Boolean = false
 
-    abstract suspend fun execute()
+    suspend fun execute() {
+        if (!isCancelled()) {
+            doExecute()
+        }
+    }
+
+    abstract suspend fun doExecute()
 
     abstract fun onTaskFailed()