+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/room/send/queue/RedactQueuedTask.kt
@@ -33,7 +33,7 @@ internal class RedactQueuedTask(
 
     override fun toString() = "[RedactQueuedTask $redactionLocalEchoId]"
 
-    override suspend fun execute() {
+    override suspend fun doExecute() {
         redactEventTask.execute(RedactEventTask.Params(redactionLocalEchoId, roomId, toRedactEventId, reason))
     }