+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/session/room/send/queue/SendEventQueuedTask.kt
@@ -35,7 +35,7 @@ internal class SendEventQueuedTask(
 
     override fun toString() = "[SendEventQueuedTask ${event.eventId}]"
 
-    override suspend fun execute() {
+    override suspend fun doExecute() {
         sendEventTask.execute(SendEventTask.Params(event, encrypt))
     }